chore: rename error handling utils and refactor imports
