+++ src/utils/__init__.py
@@ -1,6 +1,6 @@
 from .aliases import ExtensionsType, PathLike
 from .common import read_config, read_prompt, update_config
-from .error_handling import (
+from .error_handlers import (
     check_dir_path,
     check_extension,
     check_file_path,
